backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-filter>=23.5
django-cors-headers>=4.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
pathspec==0.11.2

# RAG Dependencies
sentence-transformers>=2.2.2
chromadb>=0.4.22
PyPDF2>=3.0.1
pdfplumber>=0.10.3
chardet>=5.2.0

# Celery & Redis for background tasks
celery>=5.3.4
redis>=5.0.1
django-celery-results>=2.5.1 